requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
Django==5.1.3
django-cors-headers==4.7.0
djangorestframework==3.15.2
filelock==3.17.0
fonttools==4.55.2
geopandas==1.0.1
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.3
networkx==3.4.2
numpy==2.2.0
opencv-python==4.11.0.86
osmnx==2.0.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
polyline==2.0.2
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scipy==1.15.1
shapely==2.0.6
six==1.17.0
spicy==0.16.0
sqlparse==0.5.2
svgwrite==1.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.29.3
